added ember2 and native
Ember2 is a mixture of my original function and Trav's, simply choosing which version is more efficient based on the pad size. It seems to be the fastest overall when you average out the noise.

And it seems helpful to have the native string.padStart implementation in there, especially since some of these are faster either overall or at various points.